feat(TodoAPP): Add handle keydown enter when add todo
