--- src/components/atoms/EmptyState/emptystate.style.js
@@ -1,9 +1,0 @@
-import { styled } from '@stitches/react';
-
-const WrapperEmptyState = styled('span', {
-  textAlign: 'center',
-  padding: '22px',
-  display: 'block',
-});
-
-export { WrapperEmptyState };
