--- src/components/atoms/Footer/footer.style.js
@@ -1,15 +1,0 @@
-import { styled } from '@stitches/react';
-
-const FooterTodo = styled('div', {
-  borderTop: '1px solid #87b6b8',
-  display: 'flex',
-  padding: '10px',
-  position: 'absolute',
-  bottom: '0',
-  left: '0',
-  right: '0',
-  backgroundColor: '#87B6B8',
-  fontSize: '1.3rem',
-});
-
-export { FooterTodo };
