+++ src/components/atoms/Icon/index.js
@@ -1,6 +1,13 @@
 import React from 'react';
 import Icon from '@ant-design/icons';
 
+/**
+ *
+ * @param {{
+ * component: React.ReactNode
+ * }} props
+ * @returns
+ */
 const IconComponent = (props) => {
   const { component } = props;
   return <Icon component={component} style={{ fontSize: '1.5rem' }} />;
