+++ src/components/atoms/Input/index.js
@@ -1,14 +1,33 @@
 import React from 'react';
-
 import { Input } from 'antd';
-
 import { styled } from '@stitches/react';
 
+/**
+ *
+ * @param {{
+ * placeholder: string,
+ * value: string,
+ * onChange: () => {}
+ * }} props
+ * @returns
+ */
 const AtomInput = (props) => {
-  const { placeholder, value, onChange } = props;
+  const { placeholder, value, onChange, onSubmit } = props;
+
+  const handleKeyDown = (event) => {
+    if (event.key === 'Enter' && onChange) {
+      onChange(event);
+      onSubmit();
+    }
+  };
 
   return (
-    <InputForm placeholder={placeholder} value={value} onChange={onChange} />
+    <InputForm
+      placeholder={placeholder}
+      value={value}
+      onChange={onChange}
+      onKeyDown={(event) => handleKeyDown(event)}
+    />
   );
 };
