+++ src/components/atoms/Label/index.js
@@ -1,10 +1,11 @@
+import { styled } from '@stitches/react';
 import React from 'react';
-import { LabelSpan } from './label.style';
 
 /**
  *
  * @param {{
  * text: string
+ * disabled: boolean
  * }} props
  * @returns
  */
@@ -12,3 +13,15 @@ export const LabelComponent = (props) => {
   const { text, disabled } = props;
   return <LabelSpan disabled={disabled}>{text}</LabelSpan>;
 };
+
+const LabelSpan = styled('span', {
+  fontSize: '1.8rem',
+  variants: {
+    disabled: {
+      true: {
+        color: '#c7c1c1',
+        textDecoration: 'line-through',
+      },
+    },
+  },
+});
